+++ zulu/src/test/java/us/ihmc/zulu/ZuluFlatGroundSideSteppingTest.java
@@ -3,12 +3,10 @@ package us.ihmc.zulu;
 import controller_msgs.msg.dds.FootstepDataListMessage;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
-import us.ihmc.zulu.parameters.controller.ZuluICPControllerParameters;
 import us.ihmc.zulu.parameters.controller.ZuluWalkingControllerParameters;
 import us.ihmc.avatar.AvatarFlatGroundSideSteppingTest;
 import us.ihmc.avatar.drcRobot.DRCRobotModel;
 import us.ihmc.avatar.drcRobot.RobotTarget;
-import us.ihmc.commonWalkingControlModules.capturePoint.controller.ICPControllerParameters;
 import us.ihmc.commonWalkingControlModules.configurations.WalkingControllerParameters;
 import us.ihmc.euclid.tuple3D.Vector3D;
 import us.ihmc.humanoidRobotics.communication.packets.HumanoidMessageTools;
@@ -25,18 +23,6 @@ public class ZuluFlatGroundSideSteppingTest extends AvatarFlatGroundSideStepping
       {
          return new ZuluWalkingControllerParameters(version, target, jointMap, version.getPhysicalProperties(), getContactPointParameters())
          {
-            @Override
-            public ICPControllerParameters getICPControllerParameters()
-            {
-               return new ZuluICPControllerParameters()
-               {
-                  @Override
-                  public boolean useAngularMomentum()
-                  {
-                     return true;
-                  }
-               };
-            }
          };
       }
    };
